unify into one solution file
