--- MultiplayerShaftFix/MultiplayerShaftFix.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28010.2041
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MultiplayerShaftFix", "MultiplayerShaftFix\MultiplayerShaftFix.csproj", "{CA91B161-F0EB-49DC-A848-ED90AD7F13BD}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{CA91B161-F0EB-49DC-A848-ED90AD7F13BD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CA91B161-F0EB-49DC-A848-ED90AD7F13BD}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CA91B161-F0EB-49DC-A848-ED90AD7F13BD}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CA91B161-F0EB-49DC-A848-ED90AD7F13BD}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {28583FCD-E30D-4A9C-8F27-E7191E9DCC43}
-	EndGlobalSection
-EndGlobal
