+++ MultiplayerShaftFix/manifest.json
@@ -1,0 +1,10 @@
+﻿{
+   "Name": "Multiplayer shaft fix",
+   "Author": "Prince-Leto",
+   "Version": "1.0.0",
+   "Description": "Fix the random level issue of shafts in Multiplayer",
+   "UniqueID": "princeleto.multiplayershaftfix",
+   "EntryDll": "multiplayershaftfix.dll",
+   "MinimumApiVersion": "2.0",
+   "UpdateKeys": [ "Nexus:2922" ]
+}
