+++ StardewMods.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.28407.52
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MultiplayerShaftFix", "MultiplayerShaftFix\MultiplayerShaftFix.csproj", "{CA91B161-F0EB-49DC-A848-ED90AD7F13BD}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TwentyFourHours", "TwentyFourHours\TwentyFourHours.csproj", "{2344800B-2765-4610-AFFA-359817471A21}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{CA91B161-F0EB-49DC-A848-ED90AD7F13BD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CA91B161-F0EB-49DC-A848-ED90AD7F13BD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CA91B161-F0EB-49DC-A848-ED90AD7F13BD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CA91B161-F0EB-49DC-A848-ED90AD7F13BD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2344800B-2765-4610-AFFA-359817471A21}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2344800B-2765-4610-AFFA-359817471A21}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2344800B-2765-4610-AFFA-359817471A21}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2344800B-2765-4610-AFFA-359817471A21}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {134F8E05-C11E-42F4-AB40-0CC552348F08}
+	EndGlobalSection
+EndGlobal
