--- TwentyFourHours/TwentyFourHours.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28010.2041
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TwentyFourHours", "TwentyFourHours\TwentyFourHours.csproj", "{2344800B-2765-4610-AFFA-359817471A21}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{2344800B-2765-4610-AFFA-359817471A21}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2344800B-2765-4610-AFFA-359817471A21}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2344800B-2765-4610-AFFA-359817471A21}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2344800B-2765-4610-AFFA-359817471A21}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {8219E905-4AAA-4401-8D02-6EA18D3F5FCB}
-	EndGlobalSection
-EndGlobal
